Adding arcobjects selection sample
Overview: This code is a proof of concept for two things (1) How to
interject in ArcGIS Drawing process and draw polygons on the screen (2)
How make the selection color of polgon with fill symbol which don't
overlap labels.
In ArcMap, there is a option to change the symbol of polygon selection
but it overlaps labels because these selected polygons are drawn at
GeoSelection stage which a over the top of labeling phase
(esriViewGraphics).
If one draws the fill symbol polygon at geogrphy phase they don't over
lap labels.