Update UpdateWebMap.py
Shortened the amount of inputs required.  Also have it print out the webMapID for the user when they run this.  Might be a good base to work from for future ArcGIS Online REST API items.  Removed a lot of logic required by user and just threw it into a class for simplification.