Couple of updates to the sample :construction: :construction_worker:
added a dialog box which shows the attributes and used a text edit to
modify the attribute values

A condition statement is used to check and see if the value is 0. If it
is add the string null to the value. If a user removes the value
completely they are unable to re-select that field and edit it.

Still a work in progress :construction: :construction_worker:

Shout outs to @nohe427 and @gdhakal