How to submit async print task with hybrid security settings
The samples shows how to submit async print task when your printTask is
set to be async and the security setting of ArcGIS Server is Hybrid(User
Store: Windows Domain, Role Store: ArcGIS Server Built-in). The
printTask URL may be expired in the future. You can change the URL in
the code: MainWindow.xaml.cs, line 30